Add aging out old items, add `ResourceVersion` to item data, store KubeClient and ConfigMap in ReportBuilder
Old report items are aged out based on the last OOMKill we saw. The `too
old age` is defined in the ReportBuilder struct.

The `ResourceVersion` in each item will eventually be used to determine
whether an item should be removed from a report because its
pod-controller (Deployment) has been changed in-cluster.

The KubeClient and state ConfigMap namespace/name are stored in the
ReportBuilder struct to ease reading/writing the state ConfigMap.